requirements.txt
discord.py
python-dotenv
apscheduler
aiohttp